change direction, fixer.io is out
so, getting currency exchange data is a pain in the ass, because every provider of that data occupies a niche where they are paid for API calls. fixer.io had a nice API, and they allow up to 1000 api calls per month under their free plan, which was enough for my needs in this specific case, but they do not allow you to set a base currency for the conversions, which completely vacates the utility of this interface in our specific use case (an irc robot). I may just give up on this 'correct' approach and go back to scraping XE.com, even if that does mean that I have to use a headless browser to render their craptastic javascript interface.